requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
boto3==1.28.36
botocore==1.31.36
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
ctransformers==0.2.25
dnspython==2.4.2
email-validator==2.0.0.post2
fastapi==0.103.0
filelock==3.12.3
frozenlist==1.4.0
fsspec==2023.6.0
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
multidict==6.0.4
orjson==3.9.5
packaging==23.1
Pillow==10.0.0
py-cpuinfo==9.0.0
pydantic==2.3.0
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
runpod==1.1.3
s3transfer==0.6.2
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
tqdm==4.66.1
tqdm-loggable==0.1.4
typing_extensions==4.7.1
ujson==5.8.0
urllib3==1.26.16
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
yarl==1.9.2
